# docs/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "RedeStore API",
    "description": "API da RedeStore: autenticação, loja (produtos e pedidos) e eventos (com inscrições). Endpoints protegidos exigem o header `Authorization: Bearer <token>`, obtido em `/auth/login` ou `/auth/cadastro`. Erros seguem o formato ProblemDetails (RFC 7807): `title` traz o código do erro (ex.: `EVENTO_NAO_ENCONTRADO`) e `detail` a mensagem.",
    "version": "v1"
  },
  "paths": {
...
    "/auth/cadastro": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Cadastra um novo usuário (papel 'jovem') e já devolve o token",
        "description": "Regras: nome com 2+ caracteres, e-mail válido e senha com 8+ caracteres. E-mail duplicado retorna 409 EMAIL_EM_USO.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CadastroRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpValidationProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Autentica com e-mail e senha e devolve o token JWT",
        "description": "E-mail inexistente ou senha incorreta retornam o mesmo 401 CREDENCIAIS_INVALIDAS.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpValidationProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/auth/me": {
      "get": {
        "tags": [
          "Auth"
        ],
        "summary": "Retorna os dados do usuário autenticado",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UsuarioDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
    "/auth/perfil": {
      "patch": {
        "tags": [
          "Auth"
        ],
        "summary": "Atualiza parcialmente o perfil do usuário autenticado",
        "description": "Todos os campos são opcionais; só os enviados (não nulos) são alterados. Trocar para um e-mail já usado retorna 409 EMAIL_EM_USO.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AtualizarPerfilRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UsuarioDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpValidationProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
    "/auth/recuperar-senha": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Envia por e-mail o link de redefinição de senha",
        "description": "Sempre retorna 204, exista ou não o e-mail (evita enumeração de usuários). O link aponta para Frontend:ResetPasswordUrl?token=... e expira em 1 hora.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecuperarSenhaRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "No Content"
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpValidationProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/auth/redefinir-senha": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Redefine a senha usando o token recebido por e-mail",
        "description": "O token é de uso único e expira em 1 hora. Token inválido, expirado ou já usado retorna 400 TOKEN_INVALIDO.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RedefinirSenhaRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "No Content"
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AtualizarPerfilRequest": {
        "required": [
          "nome",
          "email",
          "telefone"
        ],
        "type": "object",
        "properties": {
          "nome": {
            "type": [
              "null",
              "string"
            ]
          },
          "email": {
            "type": [
              "null",
              "string"
            ]
          },
          "telefone": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "AuthResponse": {
        "required": [
          "usuario",
          "token"
        ],
        "type": "object",
        "properties": {
          "usuario": {
            "$ref": "#/components/schemas/UsuarioDto"
          },
          "token": {
            "type": "string"
          }
        }
      },
      "CadastroRequest": {
        "required": [
          "nome",
          "email",
          "senha"
        ],
        "type": "object",
        "properties": {
          "nome": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "senha": {
            "type": "string"
          }
        }
      },
...
      "HttpValidationProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          },
          "errors": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        }
      },
...
      "LoginRequest": {
        "required": [
          "email",
          "senha"
        ],
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "senha": {
            "type": "string"
          }
        }
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "RecuperarSenhaRequest": {
        "required": [
          "email"
        ],
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          }
        }
      },
      "RedefinirSenhaRequest": {
        "required": [
          "token",
          "novaSenha"
        ],
        "type": "object",
        "properties": {
          "token": {
            "type": "string"
          },
          "novaSenha": {
            "type": "string"
          }
        }
      },
...
      "UsuarioDto": {
        "required": [
          "id",
          "nome",
          "email",
          "telefone",
          "papel"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "nome": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "telefone": {
            "type": [
              "null",
              "string"
            ]
          },
          "papel": {
            "type": "string"
          }
        }
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "JWT retornado por /auth/login ou /auth/cadastro. Validade padrão: 8 horas.",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
...
}